Add parsing and transformation for a adoc
This commit adds some clean way to parse the document, and it
also adds transformation for some of the basic attributes.

As a next step, we will try to adopt this changes to a sample
standard and see how that plays well with that.
