+++ lib/coradoc/document.rb
@@ -1,0 +1,18 @@
+require "coradoc/document/title"
+require "coradoc/document/block"
+require "coradoc/document/section"
+require "coradoc/document/attribute"
+require "coradoc/document/admonition"
+require "coradoc/document/text_element"
+require "coradoc/document/author"
+require "coradoc/document/revision"
+require "coradoc/document/header"
+require "coradoc/document/bibdata"
+require "coradoc/document/paragraph"
+require "coradoc/document/table"
+require "coradoc/document/list"
+
+module Coradoc
+  module Document
+  end
+end
