+++ lib/coradoc/document/attribute.rb
@@ -1,10 +1,20 @@
 module Coradoc
-  class Document::Attribute
-    attr_reader :key, :value
+  module Document
+    class Attribute
+      attr_reader :key, :value
 
-    def initialize(key, value)
-      @key = key
-      @value = value
+      def initialize(key, value, _options = {})
+        @key = key
+        @value = value
+      end
+    end
+
+    class Glossaries
+      attr_reader :items
+
+      def initialize(items)
+        @items = items
+      end
     end
   end
 end
