+++ lib/coradoc/document/paragraph.rb
@@ -1,0 +1,11 @@
+module Coradoc
+  module Document
+    class Paragraph
+      attr_reader :content
+
+      def initialize(content, _options = {})
+        @content = content
+      end
+    end
+  end
+end
