+++ spec/coradoc/parser/asciidoc/header_spec.rb
@@ -30,7 +30,7 @@ RSpec.describe "Coradoc::Asciidoc::Header" do
 
   module Asciidoc
     class HeaderTester < Parslet::Parser
-      include Coradoc::Asciidoc::Header
+      include Coradoc::Parser::Asciidoc::Header
 
       rule(:document) { (header | any.as(:unparsed)).repeat(1) }
       root :document
